# docs/openapi.yaml
openapi: 3.0.3
info:
  title: F4ALL Sports Talent Assessment API
  description: >
    API contract for the F4ALL AI-powered sports talent assessment platform.
    The mobile application receives a provisional on-device score, while the
    backend performs server-side verification and stores the final result.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/dashboard/reviews:
    get:
      tags:
        - Dashboard
      summary: Get review queue
      description: >
        Retrieve test submissions that require official dashboard review.

      security:
        - bearerAuth: []

      parameters:
        - name: status
          in: query
          required: false
          description: Filter results by review status
          schema:
            type: string
            enum:
              - flagged
              - approved
              - rejected
              - processing

        - name: test_type
          in: query
          required: false
          description: Filter results by test type
          schema:
            type: string
            enum:
              - vertical_jump
              - sit_ups

        - name: region
          in: query
          required: false
          description: Filter results by athlete region
          schema:
            type: string

      responses:
        '200':
          description: Review queue retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReviewItem'

        '401':
          description: Authentication required

        '403':
          description: Insufficient dashboard permissions
  /api/dashboard/reviews/{result_id}:
    get:
      tags:
        - Dashboard
      summary: Get review details
      description: >
        Retrieve complete information about a test submission for official review.

      security:
        - bearerAuth: []

      parameters:
        - name: result_id
          in: path
          required: true
          description: Test result ID
          schema:
            type: string
            format: uuid

      responses:
        '200':
          description: Review details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDetails'

        '401':
          description: Authentication required

        '403':
          description: Insufficient dashboard permissions

        '404':
          description: Result not found
  /api/dashboard/reviews/{result_id}/action:
    post:
      tags:
        - Dashboard
      summary: Take review action
      description: >
        Approve, reject, or request resubmission for a test submission.

      security:
        - bearerAuth: []

      parameters:
        - name: result_id
          in: path
          required: true
          description: Test result ID
          schema:
            type: string
            format: uuid

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewActionRequest'

      responses:
        '200':
          description: Review action completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewActionResponse'

        '400':
          description: Invalid review action

        '401':
          description: Authentication required

        '403':
          description: Insufficient dashboard permissions

        '404':
          description: Result not found


# ==============================================================
# COMPONENTS
# ==============================================================

  # -------------------------------------------------------------------------
  # Sprint 7: accounts, profile, benchmarks
  # -------------------------------------------------------------------------
...
  /api/dashboard/reviews/{result_id}/pose:
    get:
      tags: [Dashboard]
      summary: The server-extracted landmarks for the skeleton overlay
      description: "Frames of {t: ms, p: [x, y, visibility] per landmark}. Region-scoped like the result."
      security: [{bearerAuth: []}]
      parameters:
        - {name: result_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200": {description: Pose sequence}
        "404": {description: No sequence, or result outside the official's region}
  /api/dashboard/leaderboard:
    get:
      tags: [Dashboard]
      summary: Top performers from approved results, best attempt per athlete
      security: [{bearerAuth: []}]
      parameters:
        - {name: test_type, in: query, required: true, schema: {type: string}}
        - {name: region, in: query, schema: {type: string}}
        - {name: gender, in: query, schema: {type: string}}
        - {name: age_min, in: query, schema: {type: integer}}
        - {name: age_max, in: query, schema: {type: integer}}
      responses:
        "200": {description: Ranked entries}
  /api/dashboard/stats:
    get:
      tags: [Dashboard]
      summary: Result counts by status, high-severity flags and SLA breaches
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Counts, scoped to the official's region}
components:
...
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT


  # ------------------------------------------------------------
  # REQUEST SCHEMAS
  # ------------------------------------------------------------
  schemas:
...
    ReviewActionRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum:
            - approved
            - rejected
            - requested_resubmission
          example: approved

        notes:
          type: string
          nullable: true
          description: Optional note from the official
          example: 'Submission verified successfully.'


  # ------------------------------------------------------------
  # RESPONSE SCHEMAS
  # ------------------------------------------------------------
...
    ReviewItem:
      type: object
      properties:
        result_id:
          type: string
          format: uuid

        athlete_id:
          type: string
          format: uuid

        athlete_name:
          type: string
          example: 'Arun Kumar'

        test_name:
          type: string
          enum:
            - vertical_jump
            - sit_ups
          example: vertical_jump

        provisional_score:
          type: number
          format: double
          nullable: true
          example: 42

        server_score:
          type: number
          format: double
          nullable: true
          example: 41

        status:
          type: string
          example: flagged
    ReviewDetails:
      type: object
      properties:
        result_id:
          type: string
          format: uuid

        athlete:
          type: object
          properties:
            id:
              type: string
              format: uuid

            name:
              type: string

            dob:
              type: string
              format: date

            gender:
              type: string

            region:
              type: string

        test:
          type: object
          properties:
            id:
              type: string
              format: uuid

            name:
              type: string
              enum:
                - vertical_jump
                - sit_ups

        provisional_score:
          type: number
          format: double
          nullable: true

        server_score:
          type: number
          format: double
          nullable: true

        final_score:
          type: number
          format: double
          nullable: true

        status:
          type: string

        video:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid

            s3_key:
              type: string

        flags:
          type: array
          items:
            type: object
            properties:
              reason:
                type: string

              severity:
                type: string
                enum:
                  - low
                  - medium
                  - high

        face_verification:
          type: object
          nullable: true
          properties:
            verification_status:
              type: string
              enum:
                - pass
                - fail
                - manual_review

            similarity_score:
              type: number
              format: double
              nullable: true
    ReviewActionResponse:
      type: object
      properties:
        result_id:
          type: string
          format: uuid

        action:
          type: string
          enum:
            - approved
            - rejected
            - requested_resubmission

        status:
          type: string
          example: approved

        message:
          type: string
          example: Review action completed successfully.
...